checksum: rename update_checksum() -> compute_checksum()
In the context of cryptographic hashes, "update" usually means feeding
another bunch of bytes to update the digest algoritm's control
structure. Since we're going to refactor the checksumming to accept
multiple files, we'll need to run such an "update" once for each
file. Hence, rename the public function update_checksum to
compute_checksum.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>