test/bootchooser: test asymmetric slot handling for barebox and uboot
We also state supporting non-symmetric updates, e.g. with one 'recovery'
and one 'system' slot.

This scenario is tested on the slot selection level, already, but it is
not yet tested if the boot selection layer handles this properly.

Here we add tests for it to see what will go wrong...

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>