src/install: prevent g_strchomp() from being called with NULL
read_file_str() may return NULL if reading the file failed.
Calling g_strchomp() with NULL however will lead to an assertion error.

This is fixed by checking the return of read_file_str() and directly
return NULL in this case. Another benefit of this will be that we can at
least print the error message that will be returned by read_file_str()
which would have been lost otherwise.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>