src/main: merge mount_point and ext_mount_point for frontend
Both in D-Bus export of 'mountpoint' as well as in the CLI we can merge
the values of ext_mount_point and mount_point as the difference is not
relevant at this place.
Instead, the value of ext_mount_point will be of more relevance and
interest for users of the D-Bus interface and command line interface.

The behavior prior to this patch was unexpected for the user as one could
not see the mountpoint property in 'rauc status' despite the slot was
actually mounted. That was also the case for the root slot.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>