src/update_handler: implement optional ext4 resizing in img_to_fs_handler
When writing an ext4 image to a partition where the image is only as
large as required to hold the data but the partition is actually larger,
the rest of the partition will be 'hidden' to the file system unless
using resize2fs on it to extend it over the entire free partition space.

This implements handling of 'resize' option in a slot configuration for
the img_to_fs_handler by making it automatically call resize2fs after
writing the image.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>