src: introduce efi-use-bootnext option to change primary-marking behavior
This adds the efi-use-bootnext option that allows to disable the default
BootNext setting for marking a slot primary as this may be inappropriate
for use cases where the BIOS already handles the slot switching
properly.

For these cases, simply sorting the slot as first element in the
bootloader is more appropriate.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>