src/signature: pass only non-empty PINs on to PKCS#11 engine
It is convenient to pass the PIN via the environment variable. In some
production signing cases the PIN should rather be entered interactively
than being passed to RAUC via the environment variable. The environment
is built nonetheless, but with an empty RAUC_PKCS11_PIN variable. This
leads to a failing login attempt with an empty string as the PIN.

Due to the fact that the PIN is never an empty string anyway do not pass
such a PIN on and thus enable being prompted interactively for a PIN in
that case.

Signed-off-by: Bastian Krause <bst@pengutronix.de>