test/install: fix test for undefined behavior in slot selection handling
For the triple-redundancy rootfs.0, rootfs.1, rootfs.2 (including their
children), we have rootfs.0 set to 'active' so far and thus both rootfs.2
and rootfs.2 being detected as 'inactive'.

For this case, there is still no real definition of when or whether to select
rootfs.1 or rootfs.2.

Prior to this, the test checked for the selected slot being rootfs.1
which was valid for the concrete implementation, but mainly depending on
the order that was given by the GHashTable (iterator) implementation.

In recent glib version this changed resulting in rootfs.2 being selected
instead.

We adapt the test for checking correctly for both cases.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>